v1.7.0

+++ Project_Template/04.Reference-guided_Assembly_Scaffolding/LRSDAY.04.Reference-guided_Assembly_Scaffolding.sh
@@ -3,18 +3,19 @@ set -e -o pipefail
 #######################################
 # load environment variables for LRSDAY
 source ./../../env.sh
-PATH=$gnuplot_dir:$hal_dir:$PATH
+PATH=$gnuplot_dir:$PATH
 
 #######################################
 # set project-specific variables
-prefix="SK1" # The file name prefix for processing sample. Please avoid the character '.' in prefix. Default = "SK1" for the testing example.
+prefix="CPG_1a" # The file name prefix (only allowing strings of alphabetical letters, numbers, and underscores) for the processing sample. Default = "CPG_1a" for the testing example.         
+
 input_assembly="./../03.Short-read-based_Assembly_Polishing/$prefix.assembly.short_read_polished.fa" # The file path of the input genome assembly.
 ref_genome_raw="./../00.Reference_Genome/S288C.ASM205763v1.fa" # The file path of the raw reference genome.
 ref_genome_noncore_masked="./../00.Reference_Genome/S288C.ASM205763v1.noncore_masked.fa" # The file path of the specially masked reference genome where subtelomeres and chromosome-ends were hard masked. When the subtelomere/chromosome-end information is unavailable for the organism that you are interested in, you can just put the path of the raw reference genome assembly here.
 chrMT_tag="chrMT" # The sequence name for the mitochondrial genome in the raw reference genome file, if there are multiple reference mitochondrial genomes that you want to check, use a single ';' to separate them. e.g. "Sc_chrMT;Sp_chrMT". Default = "chrMT".
 gap_size=5000 # The number of Ns to insert between adjacent contigs during scaffolding. Default = "5000".
-scaffolder="ragout" # The reference-based assembly scaffolder to use: "ragout" or "ragoo". Default = "ragout". If the reference genome size is large (e.g. > 100 Mb), please use "ragoo" since extra dependency is needed for "ragout" to handle large genomes.
-threads=1 # The number of threads to use. Default = "1".
+scaffolder="ragtag" # The reference-based assembly scaffolder to use: "ragout" or "ragtag". Default = "ragtag". If the reference genome size is large (e.g. > 100 Mb), please use "ragtag" since extra dependency is needed for "ragout" to handle large genomes.
+threads=8 # The number of threads to use. Default = "8".
 debug="no" # Whether to keep intermediate files for debugging. Use "yes" if prefer to keep intermediate files, otherwise use "no". Default = "no".
 
 ######################################
@@ -26,28 +27,28 @@ sed -e '/^[^>]/s/[^ATGCatgc]/N/g' $ref_genome_noncore_masked > ref_genome_noncor
 
 if [[ $scaffolder == "ragout" ]]
 then
-    echo ".references = ref_genome" > ragout.recipe.txt
-    echo ".target = $prefix" >> ragout.recipe.txt
-    echo "ref_genome.fasta = ./ref_genome_noncore_masked.fa" >> ragout.recipe.txt
-    echo "$prefix.fasta = $input_assembly" >> ragout.recipe.txt
-    echo ".naming_ref = ref_genome" >> ragout.recipe.txt
-    source $miniconda2_dir/activate $build_dir/conda_ragout_env
-    python2 $ragout_dir/ragout -o ${prefix}_ragout_out --solid-scaffolds  -t $threads  ragout.recipe.txt
-    source $miniconda2_dir/deactivate
+    echo ".references = ref_genome" > $prefix.ragout.recipe.txt
+    echo ".target = $prefix" >>$prefix.ragout.recipe.txt
+    echo "ref_genome.fasta = ./ref_genome_noncore_masked.fa" >> $prefix.ragout.recipe.txt
+    echo "$prefix.fasta = $input_assembly" >> $prefix.ragout.recipe.txt
+    echo ".naming_ref = ref_genome" >> $prefix.ragout.recipe.txt
+    source $miniconda3_dir/activate $build_dir/ragout_conda_env
+    python3 $ragout_dir/ragout -o ${prefix}_ragout_out --solid-scaffolds --overwrite -t $threads  $prefix.ragout.recipe.txt
+    source $miniconda3_dir/deactivate
     cat ./${prefix}_ragout_out/${prefix}_scaffolds.fasta | sed "s/^>chr_/>/g" > ./${prefix}_ragout_out/${prefix}_scaffolds.renamed.fasta 
     cat ./${prefix}_ragout_out/${prefix}_scaffolds.renamed.fasta ./${prefix}_ragout_out/${prefix}_unplaced.fasta > ./${prefix}_ragout_out/${prefix}.ragout.raw.fa    
     perl $LRSDAY_HOME/scripts/adjust_assembly_by_ragoutAGP.pl -i $input_assembly -p $prefix -a ./${prefix}_ragout_out/${prefix}_scaffolds.agp -g $gap_size
-    ln -s ${prefix}.ragout.fa $prefix.assembly.ref_based_scaffolded.fa
-elif [[ $scaffolder == "ragoo" ]]
+    mv ${prefix}.ragout.fa $prefix.assembly.ref_based_scaffolded.fa
+elif [[ $scaffolder == "ragtag" ]]
 then
-    source $build_dir/py3_virtualenv_ragoo/bin/activate
-    python3 $ragoo_dir/ragoo.py -t $threads -g $gap_size  -C -m $minimap2_dir/minimap2 $input_assembly ref_genome_noncore_masked.fa
-    mv ragoo_output ${prefix}_ragoo_out
-    cp ${prefix}_ragoo_out/ragoo.fasta ${prefix}.ragoo.fa
-    ln -s ${prefix}.ragoo.fa $prefix.assembly.ref_based_scaffolded.fa
+    source $miniconda3_dir/activate $build_dir/ragtag_conda_env
+    $ragtag_dir/ragtag.py scaffold -t $threads -f 500 -g $gap_size -r -u -o ${prefix}_ragtag_out ref_genome_noncore_masked.fa $input_assembly 
+    source $miniconda3_dir/deactivate
+    cat ./${prefix}_ragtag_out/ragtag.scaffold.fasta  |sed "s/_RagTag$//g" > ${prefix}.ragtag.fa
+    mv ${prefix}.ragtag.fa $prefix.assembly.ref_based_scaffolded.fa 
 else
     echo "Unrecognized scaffolder: $scaffolder";
-    echo "Please set scaffolder=\"ragout\" or \"ragoo\".";
+    echo "Please set scaffolder=\"ragout\" or \"ragtag\".";
     echo "LRSDAY Exit!";
     exit;
 fi
@@ -66,20 +67,25 @@ perl $LRSDAY_HOME/scripts/fine_tune_gnuplot.pl -i $prefix.assembly.ref_based_sca
 $gnuplot_dir/gnuplot < $prefix.assembly.ref_based_scaffolded.filter_adjust.gp
 
 # generate dotplot for the mitochondrial genome only
-perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $ref_genome_raw -l ref.chrMT.list -m normal -o ref.chrMT.fa
-perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $prefix.assembly.ref_based_scaffolded.fa -l $prefix.assembly.ref_based_scaffolded.mt_contig.list -m normal -o $prefix.assembly.ref_based_scaffolded.mt_contig.fa
-$mummer4_dir/nucmer --maxmatch --nosimplify  -p $prefix.assembly.ref_based_scaffolded.chrMT ref.chrMT.fa $prefix.assembly.ref_based_scaffolded.mt_contig.fa
-$mummer4_dir/delta-filter -m  $prefix.assembly.ref_based_scaffolded.chrMT.delta > $prefix.assembly.ref_based_scaffolded.chrMT.delta_filter
-$mummer4_dir/mummerplot --large --postscript $prefix.assembly.ref_based_scaffolded.chrMT.delta_filter -p $prefix.assembly.ref_based_scaffolded.chrMT.filter
-perl $LRSDAY_HOME/scripts/fine_tune_gnuplot.pl -i $prefix.assembly.ref_based_scaffolded.chrMT.filter.gp -o $prefix.assembly.ref_based_scaffolded.chrMT.filter_adjust.gp -r ref.chrMT.fa -q $prefix.assembly.ref_based_scaffolded.mt_contig.fa
-$gnuplot_dir/gnuplot < $prefix.assembly.ref_based_scaffolded.chrMT.filter_adjust.gp
+if [ -s $prefix.assembly.ref_based_scaffolded.mt_contig.list ]
+then
+    perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $ref_genome_raw -l ref.chrMT.list -m normal -o ref.chrMT.fa
+    perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $prefix.assembly.ref_based_scaffolded.fa -l $prefix.assembly.ref_based_scaffolded.mt_contig.list -m normal -o $prefix.assembly.ref_based_scaffolded.mt_contig.fa
+    $mummer4_dir/nucmer --maxmatch --nosimplify  -p $prefix.assembly.ref_based_scaffolded.chrMT ref.chrMT.fa $prefix.assembly.ref_based_scaffolded.mt_contig.fa
+    $mummer4_dir/delta-filter -m  $prefix.assembly.ref_based_scaffolded.chrMT.delta > $prefix.assembly.ref_based_scaffolded.chrMT.delta_filter
+    $mummer4_dir/mummerplot --large --postscript $prefix.assembly.ref_based_scaffolded.chrMT.delta_filter -p $prefix.assembly.ref_based_scaffolded.chrMT.filter
+    perl $LRSDAY_HOME/scripts/fine_tune_gnuplot.pl -i $prefix.assembly.ref_based_scaffolded.chrMT.filter.gp -o $prefix.assembly.ref_based_scaffolded.chrMT.filter_adjust.gp -r ref.chrMT.fa -q $prefix.assembly.ref_based_scaffolded.mt_contig.fa
+    $gnuplot_dir/gnuplot < $prefix.assembly.ref_based_scaffolded.chrMT.filter_adjust.gp
+else 
+    echo "no chrMT contig detected, therefore no chrMT plotting will be performed."
+fi
 
 # clean up intermediate files
 if [[ $debug == "no" ]]
 then
     if [[ $scaffolder == "ragout" ]]
     then
-	rm ragout.recipe.txt
+	rm $prefix.ragout.recipe.txt
     fi
     rm ref_genome_noncore_masked.fa
     rm *.filter.fplot
@@ -89,8 +95,11 @@ then
     rm *.filter.gp
     rm *.filter_adjust.gp
     rm *.filter.ps
-    rm ref.chrMT.list
-    rm ref.chrMT.fa
+    if [ -e ref.chrMT.fa ]
+    then
+	rm ref.chrMT.list
+	rm ref.chrMT.fa
+    fi
 fi
 
 ############################
