+++ Project_Template/11.TE_Annotation/LRSDAY.11.TE_Annotation.sh
@@ -6,24 +6,32 @@ source ./../../env.sh
 
 #######################################
 # set project-specific variables
-prefix="SK1" # The file name prefix for the processing sample. Default = "SK1" for the testing example.
-genome="./../07.Supervised_Final_Assembly/$prefix.assembly.final.fa" # The file path of the input genome assembly.
-threads=1 # The number of threads to use. Default = "1".
+prefix="CPG_1a" # The file name prefix (only allowing strings of alphabetical letters, numbers, and underscores) for the processing sample. Default = "CPG_1a" for the testing example.          
+genome_assembly="./../07.Supervised_Final_Assembly/$prefix.assembly.final.fa" # The path of the final input genome assembly.
+chrMT_tag="chrMT" # The sequence name for the mitochondrial genome in the final assembly. If there are multiple sequences, use a single ';' to separate them. e.g. "chrMT_part1;chrMT_part2". Default = "chrMT".
+threads=8 # The number of threads to use. Default = "8".
 debug="no" # Whether to keep intermediate files for debugging. Use "yes" if prefer to keep intermediate files, otherwise use "no". Default = "no".
 
 #######################################
 # process the pipeline
 
-ln -s $genome genome.fa
-$repeatmasker_dir/RepeatMasker -pa $threads -lib $LRSDAY_HOME/data/TY_lib.Yue_et_al_2017_NG.fa  genome.fa  -xsmall  -gff
-$reannotate_dir/REannotate_longname  -g -f $LRSDAY_HOME/data/fuzzy_defragmentation.txt  -k $clustalw_dir/clustalw2 -r 0.48764  -d 10000  -t genome.fa.out  genome.fa
+TE_lib="$LRSDAY_HOME/data/TE_lib.v20221221.tidy.fa"
+TE_lib_LTRonly="$LRSDAY_HOME/data/TE_lib.v20221221.LTRonly.tidy.fa"
+
+echo $chrMT_tag | sed -e "s/;/\n/g" > $prefix.assembly.chrMT.list
+perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $genome_assembly -l $prefix.assembly.chrMT.list -m reverse -o $prefix.assembly.nuclear_genome.fa
+perl $LRSDAY_HOME/scripts/tidy_fasta.pl -i $prefix.assembly.nuclear_genome.fa -o $prefix.assembly.nuclear_genome.tidy.fa
+ln -s $prefix.assembly.nuclear_genome.tidy.fa  $prefix.genome.fa
+
+$repeatmasker_dir/RepeatMasker -pa $threads -lib $TE_lib  $prefix.genome.fa  -xsmall  -gff
+$reannotate_dir/REannotate_longname  -g -f $LRSDAY_HOME/data/fuzzy_defragmentation.txt  -k $clustalw_dir/clustalw2 -r 0.48764  -d 10000  -t $prefix.genome.fa.out  $prefix.genome.fa
 mv REannotate_output ${prefix}_REannotate_out
 
 cd ${prefix}_REannotate_out 
 
 perl $LRSDAY_HOME/scripts/parse_REannotate_gff.pl -o $prefix.REannotate.raw.gff
 mv $prefix.REannotate.raw.gff $prefix.REannotate.gff
-perl $LRSDAY_HOME/scripts/parse_REannotate_out.pl -i genome.fa.REannotation -p $prefix.TY_REannotate -r ./../genome.fa
+perl $LRSDAY_HOME/scripts/parse_REannotate_out.pl -i $prefix.genome.fa.REannotation -p $prefix.TY_REannotate -r ./../$prefix.genome.fa
 
 cat $prefix.TY_REannotate.complete.raw.fa  |egrep ">" |egrep  '(TY1|TY2)' |sed "s/>//g" >$prefix.TY_REannotate.complete.TY1TY2.raw.list
 perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $prefix.TY_REannotate.complete.raw.fa -l $prefix.TY_REannotate.complete.TY1TY2.raw.list  -o $prefix.TY_REannotate.complete.TY1TY2.raw.fa
@@ -31,11 +39,12 @@ perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $prefix.TY_REannotate.compl
 cat $prefix.TY_REannotate.truncated.raw.fa  |egrep ">" |egrep  '(TY1|TY2)' |sed "s/>//g" >$prefix.TY_REannotate.truncated.TY1TY2.raw.list
 perl $LRSDAY_HOME/scripts/select_fasta_by_list.pl -i $prefix.TY_REannotate.truncated.raw.fa -l $prefix.TY_REannotate.truncated.TY1TY2.raw.list  -o $prefix.TY_REannotate.truncated.TY1TY2.raw.fa
 
-for i in {3..5}
+for i in TY3 TY4 TY5 TSU4
 do
-    cat $prefix.TY_REannotate.complete.raw.fa  |egrep ">" |egrep "TY${i}" |sed "s/>//g"  >$prefix.TY_REannotate.complete.TY${i}.final.list
-    cat $prefix.TY_REannotate.truncated.raw.fa  |egrep ">" |egrep "TY${i}" |sed "s/>//g"  >$prefix.TY_REannotate.truncated.TY${i}.final.list
+    cat $prefix.TY_REannotate.complete.raw.fa  |egrep ">" |egrep "$i" |sed "s/>//g"  >$prefix.TY_REannotate.complete.$i.final.list
+    cat $prefix.TY_REannotate.truncated.raw.fa  |egrep ">" |egrep "$i" |sed "s/>//g"  >$prefix.TY_REannotate.truncated.$i.final.list
 done
+
 
 TY2_query="$LRSDAY_HOME/data/TY2_specific_region.fa"
 
@@ -74,25 +83,25 @@ else
     cat $prefix.TY_REannotate.truncated.TY2.list |sed "s/TY1/TY2/g" >$prefix.TY_REannotate.truncated.TY2.final.list
 fi
 
-db="$LRSDAY_HOME/data/TY_lib.Yue_et_al_2017_NG.LTRonly.fa"
+LTR_query="$prefix.TY_REannotate.soloLTR.raw.fa"
+db=$TE_lib_LTRonly
 db_tag="soloLTR_db";
-if [ ! -s "$db" ]
+if [ ! -s "$LTR_query" ]
 then
-    echo "$db is empty, skip .."
+    echo "$LTR_query is empty, skip .."
     touch $prefix.TY_soloLTR.refined.nr.gff
 else
-    LTR_query="$prefix.TY_REannotate.soloLTR.raw.fa"
-    $blast_dir/makeblastdb  -in $db  -dbtype nucl -title $db_tag -hash_index -out $db_tag
-    $blast_dir/blastn  -task blastn -query $LTR_query -num_threads $threads -db $db_tag -outfmt 7 >$prefix.$db_tag.soloLTR.blastn.fmt7.out
-    perl $LRSDAY_HOME/scripts/trim_soloLTR_by_blast.pl -q $LTR_query -b  $prefix.$db_tag.soloLTR.blastn.fmt7.out -p $prefix  -i 70 -l 100
+    $blast_dir/makeblastdb -in $db -dbtype nucl -title $db_tag -hash_index -out $db_tag
+    $blast_dir/blastn -task blastn -query $LTR_query -num_threads $threads -db $db_tag -outfmt 7 >$prefix.$db_tag.soloLTR.blastn.fmt7.out
+    perl $LRSDAY_HOME/scripts/trim_soloLTR_by_blast.pl -q $LTR_query -b  $prefix.$db_tag.soloLTR.blastn.fmt7.out -p $prefix  -i 75 -l 100
     $bedtools_dir/bedtools sort  -i  $prefix.TY_REannotate.soloLTR.refined.gff > $prefix.TY_REannotate.soloLTR.refined.sorted.gff
-    perl $LRSDAY_HOME/scripts/rm_overlap_features_from_gff_simple.pl  -r ./../genome.fa  -i $prefix.TY_REannotate.soloLTR.refined.sorted.gff -o $prefix.TY_soloLTR.refined.nr.gff 
+    perl $LRSDAY_HOME/scripts/rm_overlap_features_from_gff_simple.pl  -r ./../$prefix.genome.fa  -i $prefix.TY_REannotate.soloLTR.refined.sorted.gff -o $prefix.TY_soloLTR.refined.nr.gff 
 fi
 
-for i in {1..5}
+for i in TY1 TY2 TY3 TY4 TY5 TSU4
 do
-    perl $LRSDAY_HOME/scripts/TY_list2gff3.pl -i $prefix.TY_REannotate.complete.TY${i}.final.list -o $prefix.TY_REannotate.complete.TY${i}.final.gff -t $prefix
-    perl $LRSDAY_HOME/scripts/TY_list2gff3.pl -i $prefix.TY_REannotate.truncated.TY${i}.final.list -o $prefix.TY_REannotate.truncated.TY${i}.final.gff -t $prefix
+    perl $LRSDAY_HOME/scripts/TY_list2gff3.pl -i $prefix.TY_REannotate.complete.$i.final.list -o $prefix.TY_REannotate.complete.$i.final.gff -t $prefix
+    perl $LRSDAY_HOME/scripts/TY_list2gff3.pl -i $prefix.TY_REannotate.truncated.$i.final.list -o $prefix.TY_REannotate.truncated.$i.final.gff -t $prefix
 done
 
 cat $prefix.TY_REannotate.*.final.gff > $prefix.TY.complete_plus_truncated.final.gff
@@ -102,17 +111,22 @@ $bedtools_dir/bedtools intersect -v -a $prefix.TY_soloLTR.refined.nr.gff -b $pre
 
 cat $prefix.TY.complete_plus_truncated.final.gff $prefix.TY.soloLTR.final.gff > $prefix.TY.all.final.gff
 
-perl $LRSDAY_HOME/scripts/tidy_maker_gff3.pl -r ./../genome.fa -i $prefix.TY.all.final.gff -o $prefix.TE.gff3 -t $prefix 
+#perl $LRSDAY_HOME/scripts/tidy_maker_gff3.pl -r ./../$prefix.genome.fa -i $prefix.TY.all.final.gff -o $prefix.TE.gff3 -t $prefix 
+perl $LRSDAY_HOME/scripts/tidy_TE_gff3.pl -r ./../$prefix.genome.fa -i $prefix.TY.all.final.gff -o ./../$prefix.nuclear_genome.TE.gff3 -t $prefix 
 
-cp $prefix.TE.gff3 ./../
-cp $prefix.REannotate.gff ./../
-rm $prefix.*.final.gff
+
+if [[ $debug == "no" ]]
+then
+    rm $prefix.*.final.gff
+fi
+
 cd ..
 
 # clean up intermediate files
 if [[ $debug == "no" ]]
 then
-    rm genome.fa*
+    rm $prefix.assembly.nuclear_genome*
+    rm $prefix.genome.fa*
 fi
 
 ############################
